+++ system/Template/Factory.php
@@ -3,7 +3,7 @@
 namespace Template;
 
 use Core\Config;
-use Core\Language;
+use Language\Language;
 use Support\Contracts\ArrayableInterface as Arrayable;
 use Template\Template;
 use View\Factory as ViewFactory;
